pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ppiconf"
version = "0.1.0"
description = "Deterministic configuration for Dockerized apps"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.0"
]

[project.scripts]
ppiconf-gen = "ppiconf.main:cli_generate" # Ovo pravi CLI komandu